requirements.txt
requests==2.28.2
beautifulsoup4==4.12.2
pandas==2.0.1
scikit-learn==1.2.2
nltk==3.8.1
pyyaml==6.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
apache-airflow-providers-postgres==5.4.0
streamlit>=1.24.0
